Create an opinionated R package

Creates a new package with usethis::create_package() and applies an opinionated setup:

Usage

bootstrap_pkg(
  path,
  author_name = "Bjarke Hautop Kristensen",
  author_email = "bjarke.hautop@gmail.com",
  license = usethis::use_mit_license(),
  private = FALSE
)

Arguments

path

Path where the new package is created. The last component of the path is used as the package name.

author_name

Name of the package author (aut/cre in Authors@R and the license copyright holder).

author_email

Email of the package author.

license

A call to a usethis license function, e.g. usethis::use_mit_license() or usethis::use_gpl3_license().

private

Should the GitHub repository be private?

Value

The path to the new package, invisibly. The new package is left as the active usethis project and activated with usethis::proj_activate().

Details

air formatting (usethis::use_air()), plus panache.toml and jarl.toml configs

git, pushed to GitHub (usethis::use_git(), usethis::use_github())

a license, MIT by default

an Rmd README (usethis::use_readme_rmd())

codecov test coverage with an opinionated codecov.yml (usethis::use_coverage())

spell checking (usethis::use_spell_check())

R-CMD-check and test-coverage GitHub Actions workflows (usethis::use_github_action())

Dependabot updates for GitHub Actions

a pkgdown site deployed via GitHub Pages (usethis::use_pkgdown_github_pages())

a .pre-commit-config.yaml with panache, air and jarl hooks pinned to their latest release, installed with prek if available

It finishes by running the hooks once (letting the formatters fix up the scaffolded files), re-knitting the README, and committing and pushing the result.